refactor: adjust scaffolding
