+++ src/utils.js
@@ -1,4 +1,4 @@
-export default (map) => {
+export const objectToArray = (map) => {
   return Object.keys(map).reduce((acc, key) => {
     return map[key] ? [...acc, key] : acc;
   }, []);
