+++ vite.config.js
@@ -4,11 +4,19 @@ import vue from "@vitejs/plugin-vue";
 
 export default defineConfig({
   plugins: [vue()],
+
+  resolve: {
+    alias: {
+      "@": resolve(__dirname, "./src"),
+    },
+  },
+
   build: {
     lib: {
       entry: resolve(__dirname, "./src/index.js"),
       name: "VueProgressiveImage",
     },
+
     rollupOptions: {
       external: ["vue"],
       output: {
@@ -18,6 +26,7 @@ export default defineConfig({
       },
     },
   },
+
   define: {
     VERSION: JSON.stringify(require("./package.json").version),
   },
